requirements.txt
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.2.0
tensorflow==2.13.0
prophet==1.1.4
plotly==5.13.0
jinja2==3.1.2
python-telegram-bot==20.3
yfinance==0.2.18
matplotlib==3.7.1